modularize update.sh, make it prints nicer output to sdtout

+++ update.sh
@@ -1,17 +1,60 @@
+#!/bin/bash
 
-# update ZSH configuration
-cp zsh/_config $HOME/.zshrc
+####################### Update Function ##########################
+# git configuration
+update_git()
+{
+    if [ -f $HOME/.gitconfig ] && [ -f $HOME/.gitignore_global ]
+    then
+        echo "update git configuration..."
+        cp git/_config $HOME/.gitconfig
+        cp git/_ignore $HOME/.gitignore_global
+        echo ""
+    fi
+}
 
-# update EMACS configuration
-cp -r emacs/* $HOME/.emacs.d
+# zsh configuration
+update_zsh()
+{
+    if [ -d $HOME/.oh-my-zsh ] && [ -f $HOME/.zshrc ]
+    then
+        echo "update zsh configuration..."
+        cp zsh/_config $HOME/.zshrc
+        echo ""
+    fi
+}
 
-# update VIM configuration
-cp vim/_config $HOME/.vimrc	
-vim +BundleInstall +qall
+# emacs configuration
+update_emacs()
+{
+    if [ -d $HOME/.emacs.d ]
+    then
+        echo "update emacs configuration..."
+        cp -r emacs/* $HOME/.emacs.d
+        echo ""
+    fi
+}
 
-# update Git configuration
-cp git/_config $HOME/.gitconfig
-cp git/_ignore $HOME/.gitignore_global
+# terminal configuration
+update_terminal()
+{
+    if [ -d $HOME/.config/xfce4/terminal ]
+    then
+        echo "update xfce4-terminal configuration..."
+        cp terminal/_config $HOME/.config/xfce4/terminal/terminalrc
+        echo ""
+    fi
 
-# update Terminal configuration
-cp terminal/_config $HOME/.config/xfce4/terminal/terminalrc
+}
+
+########################### Main Function ############################
+echo -e "\n=================== UPDATE ========================="
+update_git
+update_emacs
+update_zsh
+update_terminal
+
+# finishing message
+echo "Update finished. Press <enter> to continue..."
+read end
+exit 0
